refactoring code and fix brain calc game
Delete defalt value form parameter conditionsOfGame  in index.js
Destructuring array which get from gameData() in index.js
Remove getAnswer function  in index.js
removed nesting from defalt function in index.js
rename all files from games derectory remove "logic"
create file getRundomNum.js with most used function
change name of function and variables in games/brain-calc.js
fix game brain-calc: correct convert to string
