+++ bin/brain-calc.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { startGame } from '../src/games/brain-calc-logic.js';
+import { startGame } from '../src/games/brain-calc.js';
 
 startGame();
