+++ bin/brain-even.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { startGame } from '../src/games/brain-even-logic.js';
+import { startGame } from '../src/games/brain-even.js';
 
 startGame();
