+++ bin/brain-gcd.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { startGame } from '../src/games/brain-gcd-logic.js';
+import { startGame } from '../src/games/brain-gcd.js';
 
 startGame();
