+++ bin/brain-prime.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { startGame } from '../src/games/brain-prime-logic.js';
+import { startGame } from '../src/games/brain-prime.js';
 
 startGame();
