+++ bin/brain-progression.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { startGame } from '../src/games/brain-progression-logic.js';
+import { startGame } from '../src/games/brain-progression.js';
 
 startGame();
