+++ src/getRundomNum.js
@@ -1,0 +1,1 @@
+export default (value) => Math.floor(Math.random() * value);
